Port to Minecraft 26.2 + rendering refactor (v1.1.0)
  Port mod from Minecraft 1.21.10 to 26.2

  Build & Dependencies:
  - Modulus dependency updated to 1.1.0
  - Replace modImplementation/modApi with standard Gradle configurations
  - Remove Parchment mappings

  Rendering changes:
  - Add LaserRenderTypes (dedicated render type management)
  - Replace Indigo/TerrainRenderContext mixins with ModelBlockRendererMixin
    and SectionCompilerMixin for X-Ray/translucent block rendering
  - Remove Sodium-specific mixins (MixinBlockOcclusionCache,
    MixinBlockRenderer) — no longer needed with new approach

  Removed:
  - TBSuperpower (merged into Modulus base classes)
  - IHasOverlay interface (functionality moved to Modulus)
  - EntitySavingFields, IIndigoAlphaContext interfaces
  - TBCommonUtil (logic moved to Modulus CommonUtil)
  - BlockRendererMixin, SimpleBlockRenderContextMixin,
    TerrainRenderContextMixin (replaced by new mixins)

  Refactors:
  - Simplify FlightAbility and SpeedAbility (delegate to Modulus base)
  - Reduce EntityMixin and LivingEntityMixin (logic moved to Modulus)
  - Update all abilities for new Modulus API
  - Update mixin targets for 26.2 API changes

+++ src/main/java/chappie/theboys/common/particle/WaterSplashParticle.java
@@ -38,9 +38,9 @@ public class WaterSplashParticle extends SingleQuadParticle {
                 this.zd *= 1.1;
             }
 
-            this.xd *= (double)this.friction;
-            this.yd *= (double)this.friction;
-            this.zd *= (double)this.friction;
+            this.xd *= this.friction;
+            this.yd *= this.friction;
+            this.zd *= this.friction;
             if (this.onGround) {
                 this.xd *= 0.7F;
                 this.zd *= 0.7F;
